# docs/swagger.yaml
basePath: /
definitions:
  controllers.Response-response_Customer:
    properties:
      data:
        $ref: '#/definitions/response.Customer'
      message:
        type: string
      status:
        type: string
    type: object
  controllers.Response-string:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  request.Customer:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  response.Customer:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Berikut API Loyalty Point Agent.
  termsOfService: http://swagger.io/terms/
  title: API Loyalty Point Agent
  version: "1.0"
paths:
  /customer/login:
    post:
      consumes:
      - application/json
      description: Log in a customer with the given email and password
      parameters:
      - description: example value for login; email = admin@example.com, name = admin,
          password = admin123
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/request.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/controllers.Response-string'
        "201":
          description: success
          schema:
            $ref: '#/definitions/controllers.Response-string'
        "400":
          description: failed
          schema:
            $ref: '#/definitions/controllers.Response-string'
        "401":
          description: failed
          schema:
            $ref: '#/definitions/controllers.Response-string'
      summary: Log in a customer
      tags:
      - customers
  /customer/register:
    post:
      consumes:
      - application/json
      description: Register a new customer with the given details
      parameters:
      - description: example value for registration; email = admin@example.com, name
          = admin, password = admin123
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/request.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/controllers.Response-response_Customer'
        "201":
          description: success
          schema:
            $ref: '#/definitions/controllers.Response-response_Customer'
        "400":
          description: failed
          schema:
            $ref: '#/definitions/controllers.Response-string'
      summary: Register a new customer
      tags:
      - customers
swagger: "2.0"